secrets-manager-reader/pyproject.toml
[project]
name = "secrets-manager-reader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-auth>=2.48.0",
    "google-cloud-secret-manager>=2.26.0",
]
